adlaan-contract-agent/requirements.txt
fastapi
uvicorn[standard]

# LangGraph - pinned to latest versions (August 2025)
langgraph==0.6.5
langchain-core==0.3.74
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.14

# Google AI integration
langchain-google-genai

# Environment loading
python-dotenv
